requirements.txt
Click==8.1.3
Jinja2==3.1.2
PyYAML==6.0
inquirer==2.10.1
typing_extensions==4.5.0
ruamel.yaml==0.17.2
requests==2.28.2
pycryptodome==3.17
rlp==3.0.0
py-solc-x==1.1.1
cbor2==5.4.6
toml==0.10.2
pydantic[dotenv]~=1.10
